requirements.txt
accelerate==0.34.2
brotlipy==0.7.0
codewithgpu==0.2.8
conda==22.11.1
conda-content-trust==0.1.3
conda-package-handling==1.9.0
distro==1.9.0
einops==0.8.0
Flask==3.0.3
fqdn==1.5.1
gradio==4.44.1
gunicorn==23.0.0
ipywidgets==8.1.1
isoduration==20.11.0
jsonpointer==2.4
jupyterlab==4.0.10
jupyterlab-language-pack-zh-CN==4.0.post6
loguru==0.7.2
openai==0.28.0
pip==22.3.1
pyOpenSSL==22.0.0
PySocks==1.7.1
scipy==1.14.1
supervisor==4.2.5
tensorboard==2.15.1
tiktoken==0.8.0
toolz==0.12.0
torchvision==0.16.2+cu121
transformers-stream-generator==0.0.4
uri-template==1.3.0
webcolors==1.13
wheel==0.37.1
